feat: bind the thumbnail and trigger media upload task
