+++ MastodonSDK/Sources/MastodonAsset/Assets.xcassets/Scene/Compose/Attachment/retry.imageset/Arrow Clockwise.pdf
@@ -1,0 +1,91 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 2.750000 2.750000 cm
+0.000000 0.000000 0.000000 scn
+9.250000 16.500000 m
+5.245935 16.500000 2.000000 13.254065 2.000000 9.250000 c
+2.000000 5.245935 5.245935 2.000000 9.250000 2.000000 c
+13.254065 2.000000 16.500000 5.245935 16.500000 9.250000 c
+16.500000 9.535608 16.483484 9.817360 16.451357 10.094351 c
+16.383255 10.681498 16.809317 11.250000 17.400400 11.250000 c
+17.916018 11.250000 18.369314 10.891933 18.431660 10.380100 c
+18.476776 10.009713 18.500000 9.632568 18.500000 9.250000 c
+18.500000 4.141366 14.358634 0.000000 9.250000 0.000000 c
+4.141366 0.000000 0.000000 4.141366 0.000000 9.250000 c
+0.000000 14.358634 4.141366 18.500000 9.250000 18.500000 c
+11.423139 18.500000 13.421247 17.750608 15.000000 16.496151 c
+15.000000 17.000000 l
+15.000000 17.552284 15.447716 18.000000 16.000000 18.000000 c
+16.552284 18.000000 17.000000 17.552284 17.000000 17.000000 c
+17.000000 14.301708 l
+17.011232 14.284512 17.022409 14.267276 17.033529 14.250000 c
+17.000000 14.250000 l
+17.000000 14.000000 l
+17.000000 13.447716 16.552284 13.000000 16.000000 13.000000 c
+13.000000 13.000000 l
+12.447715 13.000000 12.000000 13.447716 12.000000 14.000000 c
+12.000000 14.552284 12.447715 15.000000 13.000000 15.000000 c
+13.666476 15.000000 l
+12.443584 15.940684 10.912110 16.500000 9.250000 16.500000 c
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  1365
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 24.000000 24.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Pages 5 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000001455 00000 n
+0000001478 00000 n
+0000001651 00000 n
+0000001725 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+1784
+%%EOF
