+++ MastodonSDK/Sources/MastodonAsset/Assets.xcassets/Scene/Compose/Attachment/stop.imageset/Dismiss.pdf
@@ -1,0 +1,89 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 4.000000 3.804749 cm
+0.000000 0.000000 0.000000 scn
+0.209704 15.808150 m
+0.292893 15.902358 l
+0.653377 16.262842 1.220608 16.290571 1.612899 15.985547 c
+1.707107 15.902358 l
+8.000000 9.610251 l
+14.292892 15.902358 l
+14.683416 16.292883 15.316584 16.292883 15.707108 15.902358 c
+16.097631 15.511834 16.097631 14.878669 15.707108 14.488145 c
+9.415000 8.195251 l
+15.707108 1.902359 l
+16.067591 1.541875 16.095320 0.974643 15.790295 0.582352 c
+15.707108 0.488144 l
+15.346623 0.127661 14.779391 0.099932 14.387100 0.404957 c
+14.292892 0.488144 l
+8.000000 6.780252 l
+1.707107 0.488144 l
+1.316582 0.097620 0.683418 0.097620 0.292893 0.488144 c
+-0.097631 0.878668 -0.097631 1.511835 0.292893 1.902359 c
+6.585000 8.195251 l
+0.292893 14.488145 l
+-0.067591 14.848629 -0.095320 15.415859 0.209704 15.808150 c
+0.292893 15.902358 l
+0.209704 15.808150 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  914
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 24.000000 24.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Pages 5 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000001004 00000 n
+0000001026 00000 n
+0000001199 00000 n
+0000001273 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+1332
+%%EOF
