+++ MastodonSDK/Sources/MastodonSDK/API/Mastodon+API+V2+Media.swift
@@ -43,6 +43,20 @@ extension Mastodon.API.V2.Media {
         request.timeoutInterval = 180    // should > 200 Kb/s for 40 MiB media attachment
         let serialStream = query.serialStream
         request.httpBodyStream = serialStream.boundStreams.input
+        
+        // total unit count in bytes count
+        // will small than actally count due to multipart protocol meta
+        serialStream.progress.totalUnitCount = {
+            var size = 0
+            size += query.file?.sizeInByte ?? 0
+            size += query.thumbnail?.sizeInByte ?? 0
+            return Int64(size)
+        }()
+        query.progress.addChild(
+            serialStream.progress,
+            withPendingUnitCount: query.progress.totalUnitCount
+        )
+        
         return session.dataTaskPublisher(for: request)
             .tryMap { data, response in
                 let value = try Mastodon.API.decode(type: Mastodon.Entity.Attachment.self, from: data, response: response)
